+++ lib/main.dart
@@ -1,8 +1,8 @@
 import "package:flutter/material.dart";
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:get/get.dart';
-import 'package:ilkda_client/view/00_login_page/login_page.dart';
-import 'package:ilkda_client/view/10_home_page/home_page.dart';
+import 'package:ilkda_client/view/0_login_page/login_page.dart';
+import 'package:ilkda_client/view/1_home_page/home_page.dart';
 import 'package:ilkda_client/view_model/home_page_viewcontroller.dart';
 import 'package:ilkda_client/view_model/user_controller.dart';
 import 'package:kakao_flutter_sdk_user/kakao_flutter_sdk_user.dart';
@@ -23,7 +23,7 @@ class IlkdaClient extends StatelessWidget {
         return GetMaterialApp(
           debugShowCheckedModeBanner: false,
           title: '읽다',
-          initialRoute: "/LogIn",
+          initialRoute: "/Home",
           initialBinding: BindingsBuilder(() {}),
           getPages: [
             GetPage(
