+++ lib/model/aladin_book.dart
